refactor: 딜러에서 플레이어를 받아서 수익률을 계산하도록 수정
